refactor, deleted unnecessary || unused files
