+++ lib/app/presentation/modules/recipes/widgets/recipe_card/recipe_card_actions_row.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:groceries/app/presentation/modules/recipes/widgets/bottom_sheets/add_recipe_to_list_bottom_sheet.dart';
 import 'package:groceries/app/presentation/modules/recipes/widgets/bottom_sheets/delete_recipe_bottom_sheet.dart';
 import 'package:groceries/app/presentation/modules/recipes/widgets/recipe_card/recipe_card_action_button.dart';
-import 'package:groceries/app/presentation/widgets/constant_widgets/app_snack_bars.dart';
+import 'package:groceries/app/presentation/widgets/app_snack_bars.dart';
 import 'package:groceries/gen/assets.gen.dart';
 
 class RecipeCardActionsRow extends StatelessWidget {
