--- lib/app/presentation/widgets/buttons/app_transparent_button.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppTransparentButton extends StatelessWidget {
-  const AppTransparentButton({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
